Title: 두수의 곱과 몫, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240311/두수의 곱과 몫/multiple-and-quotient-of-two-numbers.cpp
@@ -6,7 +6,7 @@ int main() {
     int a ,b ;
     cin >> a >> b ; 
     cout << a << " * "<< b << " = " << a*b <<endl;
-    cout << a << " / "<< b << " = " << (double)a/b <<endl;
+    cout << a << " / "<< b << " = " << a/b <<endl;
 
     // 여기에 코드를 작성해주세요.
     return 0;
